canonical format check of enums for runtime config
